fix all the configs and get things working again
