+++ package.json
@@ -6,9 +6,12 @@
     "dev": "next dev src/ui",
     "build": "next build src/ui",
     "lint": "next lint src/ui",
-    "type-check": "tsc -p src/ui/tsconfig.json --noEmit",
+    "type-check": "tsc -p src/ui/tsconfig.json --noEmit && tsc -p src/ui/tsconfig.test.json --noEmit && tsc -p src/ui/tsconfig.cypress.json --noEmit",
     "format": "prettier --ignore-path src/ui/.prettierignore --write \"src/ui/**/*.+(js|ts|jsx|tsx|json)\"",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "test": "jest --config src/ui/jest.config.js --passWithNoTests src/ui",
+    "test:unit": "jest --config src/ui/jest.config.js --passWithNoTests src/ui/tests/unit",
+    "test:integration": "jest --config src/ui/jest.config.js --passWithNoTests src/ui/tests/integration"
   },
   "dependencies": {
     "next": "15.3.2",
@@ -16,23 +19,36 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
     "@eslint/eslintrc": "^3",
     "@types/node": "^22",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
+    "@types/testing-library__jest-dom": "^5.14.9",
     "@typescript-eslint/eslint-plugin": "^8.33.1",
     "@typescript-eslint/parser": "^8.33.1",
+    "cypress": "^13.13.3",
     "eslint": "^9.0.0",
     "eslint-config-next": "15.3.2",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-no-secrets": "^1.0.2",
     "eslint-plugin-prettier": "^5.4.0",
     "eslint-plugin-react": "^7.31.10",
     "eslint-plugin-react-hooks": "^5.2.0",
     "husky": "^9.1.7",
+    "jest": "^29.7.0",
+    "jest-coverage-badges": "^1.1.2",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-runner-groups": "^2.2.0",
+    "jest-transform-stub": "^2.0.0",
     "prettier": "^3.5.3",
+    "ts-jest": "^29.2.4",
     "typescript": "^5"
   },
   "lint-staged": {
