+++ src/ui/jest.config.js
@@ -1,23 +1,20 @@
 const nextJest = require('next/jest');
 
 const createJestConfig = nextJest({
-  dir: './',
+  dir: __dirname,
 });
 
 const customJestConfig = {
   collectCoverage: false,
-  collectCoverageFrom: ['src'],
+  collectCoverageFrom: ['tests'],
   coverageDirectory: './.meta',
   coverageReporters: ['json-summary'],
   moduleFileExtensions: ['ts', 'tsx', 'js'],
   moduleNameMapper: {
-    '^@/(.*)$': '<rootDir>/src/$1',
+    '^@/(.*)$': '<rootDir>/$1',
   },
   setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
   testEnvironment: 'jest-environment-jsdom',
-  transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-  },
 };
 
 module.exports = createJestConfig(customJestConfig);
