+++ src/ui/tsconfig.json
@@ -1,38 +1,11 @@
 {
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
+    /* only runtime types here */
+    "types": ["node"],
     "allowJs": false,
-    "allowSyntheticDefaultImports": true,
-    "baseUrl": "src/ui",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "incremental": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "module": "esnext",
-    "moduleResolution": "node",
-    "noEmit": true,
-    "noFallthroughCasesInSwitch": true,
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ],
-    "paths": {
-      "@/*": ["./src/*"]
-    },
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "target": "ES2022"
+    "isolatedModules": true
   },
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-    ".next/types/**/*.ts",
-    "next-env.d.ts",
-    "src",
-    "app"
-  ],
-  "exclude": [".meta/**/*", ".next/**/*", ".vscode", "node_modules"]
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app"],
+  "exclude": ["**/*.test.ts", "**/*.test.tsx"]
 }
