--- ui/.eslintrc.json
@@ -1,64 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "es6": true,
-    "jest": true
-  },
-  "extends": [
-    "eslint:recommended",
-    "next",
-    "next/core-web-vitals",
-    "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:prettier/recommended"
-  ],
-  "ignorePatterns": [".vscode", "jest.config.js"],
-  "overrides": [
-    {
-      "files": ["*.mjs"],
-      "parser": "espree", // Use the default JavaScript parser for `.mjs` files
-      "parserOptions": {
-        "sourceType": "module"
-      }
-    }
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaFeatures": {
-      "jsx": true
-    },
-    "ecmaVersion": 2018,
-    "project": "./tsconfig.json",
-    "sourceType": "module"
-  },
-  "plugins": ["@typescript-eslint", "import", "jest", "no-secrets", "react"],
-  "root": true,
-  "rules": {
-    "complexity": ["warn", { "max": 8 }],
-    "curly": ["error", "all"],
-    "import/order": [
-      "error",
-      {
-        "groups": [["builtin", "external", "internal"]],
-        "newlines-between": "always"
-      }
-    ],
-    "import/no-extraneous-dependencies": [
-      "error",
-      {
-        "devDependencies": [
-          "**/*.test.ts",
-          "**/*.test.tsx",
-          "**/*.d.ts",
-          "**/*.interfaces.ts",
-          "**/*.setup.ts",
-          "**/*.config.js",
-          "**/*.config.mjs"
-        ],
-        "optionalDependencies": false,
-        "peerDependencies": false
-      }
-    ],
-    "no-secrets/no-secrets": ["error", { "additionalRegexes": {}, "ignoreContent": [] }]
-  }
-}
