--- ui/.prettierignore
@@ -1,6 +1,0 @@
-/node_modules
-.vscode
-/bin
-/build
-/public
-/.meta
