--- ui/DEVELOPING.md
@@ -1,33 +1,0 @@
-## Getting Started
-
-First, run the development server:
-
-```bash
-npm run dev
-```
-
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
-
-#### `npm run build`
-
-Builds the app for production to the `out` folder.\
-
-#### `make test-unit`
-
-Run unit test once on your local terminal.
-
-#### `make test-integration`
-
-Run integration test once on your local terminal.
-
-#### `npx cypress run --headless`
-
-Run end to end tests against localhost:3000
-
-#### `make style`
-
-Fix code styling issues.
-
-#### `make quality`
-
-Run quality eslint quality checks.
