--- ui/next.config.mjs
@@ -1,6 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  output: 'export',
-};
-
-export default nextConfig;
