--- ui/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "guidellm",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "format": "prettier --ignore-path .prettierignore --write \"**/*.+(js|ts|jsx|tsx|json)\"",
-    "prepare": "husky install",
-    "test": "jest --passWithNoTests",
-    "test:unit": "jest --passWithNoTests tests/unit",
-    "test:integration": "jest --passWithNoTests tests/integration"
-  },
-  "dependencies": {
-    "next": "15.3.2",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^16.0.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "@types/testing-library__jest-dom": "^5.14.9",
-    "@typescript-eslint/eslint-plugin": "^8.0.0",
-    "@typescript-eslint/parser": "^8.0.0",
-    "cypress": "^13.13.3",
-    "eslint": "^8.27.0",
-    "eslint-config-next": "14.2.5",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^28.11.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-no-secrets": "^1.0.2",
-    "eslint-plugin-prettier": "^5.4.0",
-    "eslint-plugin-react": "^7.31.10",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "husky": "^9.1.7",
-    "jest": "^29.7.0",
-    "jest-coverage-badges": "^1.1.2",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-runner-groups": "^2.2.0",
-    "jest-transform-stub": "^2.0.0",
-    "prettier": "^3.5.3",
-    "ts-jest": "^29.2.4",
-    "typescript": "5"
-  },
-  "lint-staged": {
-    "*.js": "eslint --cache --fix",
-    "*.ts": "eslint --cache --fix",
-    "*.jsx": "eslint --cache --fix",
-    "*.tsx": "eslint --cache --fix"
-  }
-}
