--- ui/src/app/page.tsx
@@ -1,5 +1,0 @@
-'use client';
-
-export default function Home() {
-  return <div>GuideLLM</div>;
-}
