--- ui/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "compilerOptions": {
-    /* only runtime types here */
-    "types": [
-      "node"
-    ],
-    "allowJs": true,
-    "isolatedModules": true
-  },
-  "include": [
-    "next-env.d.ts",
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    ".next/types/**/*.ts"
-  ]
-}
